+++ src/llms/helpers/middlewares/add-provider-specific-props.ts
@@ -1,13 +1,17 @@
-import type { LanguageModelV1Middleware } from 'ai'
+import type { JSONValue, LanguageModelMiddleware } from 'ai'
 
 /** Middleware to inject provider specific properties. */
-export const addProviderSpecificProps = (provider: string, props: Record<string, any>): LanguageModelV1Middleware => ({
+export const addProviderSpecificProps = (
+	provider: string,
+	props: Record<string, JSONValue>
+): LanguageModelMiddleware => ({
+	specificationVersion: 'v3',
 	transformParams: async ({ params }) => {
 		return {
 			...params,
 
-			providerMetadata: {
-				...params.providerMetadata,
+			providerOptions: {
+				...params.providerOptions,
 				[provider]: props,
 			},
 		}