+++ src/llms/providers/openai.ts
@@ -2,7 +2,8 @@ import { createOpenAI } from '@ai-sdk/openai'
 
 import { envConfig } from '../../config/index.js'
 
-export default createOpenAI({
+const openai = createOpenAI({
 	apiKey: envConfig.OPENAI_API_KEY,
-	compatibility: 'strict', // strict mode, enable when using the OpenAI API
 })
+
+export default ((modelId: Parameters<typeof openai.chat>[0]) => openai.chat(modelId)) as typeof openai.chat